+++ src/main/java/frc/robot/teleopCommands/TeleopCargoShoot.java
@@ -5,20 +5,16 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.teleopcommands;
 
 import edu.wpi.first.wpilibj.command.Command;
 import frc.robot.Robot;
-import frc.robot.subsystems.HatchArm;
 
-
-public class LowerArmCommand extends Command {
-  public LowerArmCommand() {
-    requires(Robot.hatchArm);
+public class TeleopCargoShoot extends Command {
+  public TeleopCargoShoot() {
+    // Use requires() here to declare subsystem dependencies
+    requires(Robot.cargoArm);
   }
-
-  private final double kArmLowerMax = 0.05;
-	private final double kArmLowStall = 0.00;
 
   // Called just before this Command runs the first time
   @Override
@@ -28,25 +24,13 @@ public class LowerArmCommand extends Command {
   // Called repeatedly when this Command is scheduled to run
   @Override
   protected void execute() {
-    double angle = Robot.hatchArm.getAngle();
-		double out = 0;
-		if (angle > (5 + HatchArm.startingAngle)) {
-			out = Math.cos(Math.toRadians(angle))* kArmLowerMax;
-		}
-		if (angle < Math.toRadians(-30))
-			out *= 3;
-		else if (angle > Math.toRadians(80))
-			out -= 0.2;
-		else
-			out -= 0.1;
-		Robot.hatchArm.set(out);
-
+    
   }
 
   // Make this return true when this Command no longer needs to run execute()
   @Override
   protected boolean isFinished() {
-    return Robot.hatchArm.getAngle() < (5 + HatchArm.startingAngle);
+    return false;
   }
 
   // Called once after isFinished returns true
